[update] StringBuilder -> StringBuffer [rename] 클래스 용도별ㄹ 서로 다른 패키지로 옮김
김재환 튜터님의 피드백 참고하여 프로젝트 수정하였습니다.

<<<< 김재환 튜터님 피드백>>>>

[좋았던점]
-. String으로 써져 있던 부분들 StringBuilder로 잘 수정해 주신 부분 보았고, 훌륭합니다:웃음:(혹시 멀티스레드 동기화 안될 수 있어서 멀티 스레드 환경일시에는 StringBuffer로 추천합니다.)
-. 코드 컨벤션도 잘 해주셨고 정리 잘 되어 있었습니다.
[아쉬웠던 점]
-. 파일은 많은데, model이라는 폴더 하나에만 있어서 예를 들면, display부분은 view라는 폴더에 넣는다던지, 분리를 좀 더 해놓으면, 직관적으로 접근해서 유지보수가 쉬울거 같아요.
-.  Viewer.java , InOut.java , DataBase.java , CheckVaildity.java 등
코드 중간중간에 주석으로 잔여 코드가 남아 있는것이 프로젝트의 이해의 혼란을 가져다 줄 수 있고, 리뷰가 잘 이루어 지지 않은 평가를 받을 수 있어 불필요한 부분은 삭제 부탁드립니다.
-. Readme에 프로젝트의 상세내용이 있어 좋았는데, 전반적인 흐름만 명시해줘도 좋을거 같습니다. 프로젝트 개요와 차후에는 API 명세 , DB 명세 등 이 있겠네요.
